Add security for product

+++ src/main/java/com/site/cda_demo/controller/ProduitController.java
@@ -3,13 +3,15 @@ package com.site.cda_demo.controller;
 import com.site.cda_demo.dao.ProduitDao;
 import com.site.cda_demo.model.Etat;
 import com.site.cda_demo.model.Produit;
+import com.site.cda_demo.security.AppUserDetails;
 import com.site.cda_demo.security.IsAdministrateur;
-import com.site.cda_demo.security.IsRedacteur;
 import com.site.cda_demo.security.IsUtilisateur;
+import com.site.cda_demo.security.SecurityUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,10 +22,12 @@ import java.util.Optional;
 public class ProduitController {
 
   protected ProduitDao produitDao;
+  protected SecurityUtils securityUtils;
 
   @Autowired // Donne à ProduitController le ProduitDao // Injection de dépendance
-  public ProduitController(ProduitDao produitDao) {
+  public ProduitController(ProduitDao produitDao, SecurityUtils securityUtils) {
     this.produitDao = produitDao;
+    this.securityUtils = securityUtils;
   }
 
   /**
@@ -43,7 +47,6 @@ public class ProduitController {
    * @param id l'id du produit à retourner
    * @return 404 si aucun produit trouvé, ou statut 200 avec le produit
    */
-
   @GetMapping("/produit/{id}")
   @IsAdministrateur
   public ResponseEntity<Produit> get(@PathVariable int id) {
@@ -59,7 +62,12 @@ public class ProduitController {
 
 
   @PostMapping("/produit")
-  public ResponseEntity<Produit> save(@RequestBody @Valid Produit produit) {
+  @IsAdministrateur
+  public ResponseEntity<Produit> save(
+      @RequestBody @Valid Produit produit,
+      @AuthenticationPrincipal AppUserDetails userDetails) {
+
+    produit.setCreateur(userDetails.getUtilisateur());
 
     // Si le produit reçu n'a pas d'état alors, on indique qu'il est neuf par défaut
     if (produit.getEtat() == null) {
@@ -73,13 +81,24 @@ public class ProduitController {
     return new ResponseEntity<>(produit, HttpStatus.CREATED);
   }
 
-  @IsRedacteur
+  @IsAdministrateur
   @DeleteMapping("/produit/{id}")
-  public ResponseEntity<Produit> delete(@PathVariable int id) {
+  public ResponseEntity<Produit> delete(@PathVariable int id, @AuthenticationPrincipal AppUserDetails userDetails) {
     Optional<Produit> optionalProduit = produitDao.findById(id);
 
     if (optionalProduit.isEmpty()) {
       return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+
+    // On récupère le role
+    String role = securityUtils.getRole(userDetails);
+
+    // Ici on autorise le role admin a supprimer un produit qui n'est pas à lui
+    if (!role.equals("ROLE_ADMINISTRATEUR") ||
+        // Si l'id du créateur du produit est different de l'id de la personne connectée
+        optionalProduit.get().getCreateur().getId() != userDetails.getUtilisateur().getId()) {
+      return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     produitDao.deleteById(id);
@@ -89,7 +108,11 @@ public class ProduitController {
 
   // PUT pour modifier tout l'objet
   @PutMapping("/produit/{id}")
-  public ResponseEntity<Produit> update(@PathVariable int id, @RequestBody @Valid Produit produit) {
+  @IsAdministrateur
+  public ResponseEntity<Produit> update(
+      @PathVariable int id,
+      @RequestBody @Valid Produit produit,
+      @AuthenticationPrincipal AppUserDetails userDetails) {
 
     Optional<Produit> optionalProduit = produitDao.findById(id);
 
@@ -98,9 +121,10 @@ public class ProduitController {
     }
 
     produit.setId(id);
+    produit.setCreateur(userDetails.getUtilisateur());
     produitDao.save(produit);
 
-    return new ResponseEntity<>(produit, HttpStatus.NO_CONTENT);
+    return new ResponseEntity<>(produit, HttpStatus.OK);
   }
 
   // PATCH pour modifier une partie de l'opbjet
